Renamed the MSSQL sink to MSSqlServer sink
