+++ Serilog.sln
@@ -99,6 +99,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serilog.Extras.MSOwin", "sr
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serilog.Extras.MSOwin.Tests", "test\Serilog.Extras.MSOwin.Tests\Serilog.Extras.MSOwin.Tests.csproj", "{4F81EDAE-2E06-4024-925A-E495C852BDCF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MSSQLSample", "sample\MSSQLSample\MSSQLSample.csproj", "{F9BEFBBD-0883-4AAF-AEEC-6172E890A713}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serilog.Sinks.MSSqlServer", "src\Serilog.Sinks.MSSqlServer\Serilog.Sinks.MSSqlServer.csproj", "{31C64BE9-4EF6-42D5-A811-D7D24613EEF6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -256,6 +260,14 @@ Global
 		{4F81EDAE-2E06-4024-925A-E495C852BDCF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4F81EDAE-2E06-4024-925A-E495C852BDCF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4F81EDAE-2E06-4024-925A-E495C852BDCF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F9BEFBBD-0883-4AAF-AEEC-6172E890A713}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F9BEFBBD-0883-4AAF-AEEC-6172E890A713}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F9BEFBBD-0883-4AAF-AEEC-6172E890A713}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F9BEFBBD-0883-4AAF-AEEC-6172E890A713}.Release|Any CPU.Build.0 = Release|Any CPU
+		{31C64BE9-4EF6-42D5-A811-D7D24613EEF6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{31C64BE9-4EF6-42D5-A811-D7D24613EEF6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{31C64BE9-4EF6-42D5-A811-D7D24613EEF6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{31C64BE9-4EF6-42D5-A811-D7D24613EEF6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -285,22 +297,21 @@ Global
 		{EEA8B8EB-0FE6-44C2-9D10-02B7E3DEAD06} = {037440DE-440B-4129-9F7A-09B42D00397E}
 		{830BBD7E-6068-4D97-B6BF-59051FC2B6B1} = {037440DE-440B-4129-9F7A-09B42D00397E}
 		{FDA303BC-4794-4425-903E-AE40978A31F9} = {C4B270F5-E0E9-417A-B9C7-6ACBB9DB3912}
+		{C51F89BC-9DC1-4A13-98E2-A2ED09FD9F0D} = {037440DE-440B-4129-9F7A-09B42D00397E}
 		{58E93B12-6FFF-44CC-A591-68ED0BCCB9BF} = {C4B270F5-E0E9-417A-B9C7-6ACBB9DB3912}
+		{E12881F7-B522-4E42-BCCC-4A81F42F8D8B} = {037440DE-440B-4129-9F7A-09B42D00397E}
 		{E86251C0-D4C2-4751-97F1-CA70A13E87D8} = {C4B270F5-E0E9-417A-B9C7-6ACBB9DB3912}
+		{E3264D8A-63D7-40AB-B67B-BF55E53DC9C4} = {037440DE-440B-4129-9F7A-09B42D00397E}
 		{2FCAA868-28CF-4745-BECA-8B39AD41A484} = {C4B270F5-E0E9-417A-B9C7-6ACBB9DB3912}
+		{264C91AF-6631-4E04-A2AC-E32E88055228} = {037440DE-440B-4129-9F7A-09B42D00397E}
 		{1BDCB5D7-8542-4F9F-9A0B-E52AC4D09528} = {C4B270F5-E0E9-417A-B9C7-6ACBB9DB3912}
+		{21CAF132-BBAB-41FD-A018-EB9AE54822ED} = {037440DE-440B-4129-9F7A-09B42D00397E}
 		{7C03FE5B-426F-494C-9B7E-B23F060AB5D3} = {037440DE-440B-4129-9F7A-09B42D00397E}
-		{D5648551-D19D-41E3-9FC1-E74B111EEF41} = {0D135C0C-A60B-454A-A2F4-CD74A30E04B0}
-		{CE37740B-0B12-4974-BE44-A64E85C9ED9D} = {0D135C0C-A60B-454A-A2F4-CD74A30E04B0}
+		{BFC7DA08-C46F-4605-8EAA-615A884195AD} = {C4B270F5-E0E9-417A-B9C7-6ACBB9DB3912}
+		{104D5B1E-47BE-4739-A183-08C35BF4D38D} = {037440DE-440B-4129-9F7A-09B42D00397E}
 		{FDCEBC10-F403-4DDD-8594-DE6D7DD543AF} = {037440DE-440B-4129-9F7A-09B42D00397E}
 		{4F81EDAE-2E06-4024-925A-E495C852BDCF} = {0D135C0C-A60B-454A-A2F4-CD74A30E04B0}
-		{C51F89BC-9DC1-4A13-98E2-A2ED09FD9F0D} = {037440DE-440B-4129-9F7A-09B42D00397E}
-		{E12881F7-B522-4E42-BCCC-4A81F42F8D8B} = {037440DE-440B-4129-9F7A-09B42D00397E}
-		{E3264D8A-63D7-40AB-B67B-BF55E53DC9C4} = {037440DE-440B-4129-9F7A-09B42D00397E}
-		{264C91AF-6631-4E04-A2AC-E32E88055228} = {037440DE-440B-4129-9F7A-09B42D00397E}
-		{BFC7DA08-C46F-4605-8EAA-615A884195AD} = {C4B270F5-E0E9-417A-B9C7-6ACBB9DB3912}
-		{31C64BE9-4EF6-42D5-A811-D7D24613EEF6} = {037440DE-440B-4129-9F7A-09B42D00397E}
-		{104D5B1E-47BE-4739-A183-08C35BF4D38D} = {037440DE-440B-4129-9F7A-09B42D00397E}
 		{F9BEFBBD-0883-4AAF-AEEC-6172E890A713} = {C4B270F5-E0E9-417A-B9C7-6ACBB9DB3912}
+		{31C64BE9-4EF6-42D5-A811-D7D24613EEF6} = {037440DE-440B-4129-9F7A-09B42D00397E}
 	EndGlobalSection
 EndGlobal
