+++ src/Serilog.Sinks.MSSqlServer/Serilog.Sinks.MSSqlServer.nuspec
@@ -1,21 +1,21 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
   <metadata>
-    <id>Serilog.Sinks.MSSQL</id>
+    <id>Serilog.Sinks.MSSqlServer</id>
     <version>0.1.0</version>
     <authors>Michiel van Oudheusden</authors>
-    <description>Serilog event sink that writes to a MSSQL database.</description>
+    <description>Serilog event sink that writes to a Microsoft SQL Server database.</description>
     <language>en-US</language>
     <projectUrl>http://serilog.net</projectUrl>
     <licenseUrl>http://www.apache.org/licenses/LICENSE-2.0</licenseUrl>
     <iconUrl>http://serilog.net/images/serilog-nuget.png</iconUrl>
-    <tags>serilog logging MSSQL DB</tags>
+    <tags>serilog logging MS SQL Server DB</tags>
     <dependencies>
       <dependency id="Serilog" />
     </dependencies>
   </metadata>
   <files>
-    <file src="src\Serilog.Sinks.MSSQL\bin\Release\Serilog.Sinks.MSSQL.dll" target="lib\net45" />
-    <file src="src\Serilog.Sinks.MSSQL\bin\Release\Serilog.Sinks.MSSQL.xml" target="lib\net45" />
+    <file src="src\Serilog.Sinks.MSSqlServer\bin\Release\Serilog.Sinks.MSSqlServer.dll" target="lib\net45" />
+    <file src="src\Serilog.Sinks.MSSqlServer\bin\Release\Serilog.Sinks.MSSqlServer.xml" target="lib\net45" />
   </files>
 </package>
