+++ src/Serilog.Sinks.MSSqlServer/packages.config
